# src/main/java/com/man/fota/dto/FeatureDTO.java
package com.man.fota.dto;

public class FeatureDTO {

    private String identification;

    public String getIdentification() {
        return identification;
    }

    public void setIdentification(String identification) {
        this.identification = identification;
    }
}
