# src/main/java/com/man/fota/dto/VehicleDTO.java
package com.man.fota.dto;

public class VehicleDTO {

    private String identification;

    public String getIdentification() {
        return identification;
    }

    public void setIdentification(String identification) {
        this.identification = identification;
    }
}
